Improve performance on CUDA

+++ src/cpu_backend.cpp
@@ -19,26 +19,26 @@ void scanWorkItem(
     std::vector<Match>* matches,
     const std::function<void()>& flush)
 {
-    for (std::int64_t x = item.start.x; x <= item.end.x; ++x) {
-        for (std::int64_t z = item.start.z; z <= item.end.z; ++z) {
+    for (std::int32_t x = item.start.x; x <= item.end.x; ++x) {
+        for (std::int32_t z = item.start.z; z <= item.end.z; ++z) {
             if (state->cancelRequested.load(std::memory_order_relaxed)) {
                 return;
             }
 
-            for (std::int64_t y = item.start.y; y <= item.end.y; ++y) {
+            for (std::int32_t y = item.start.y; y <= item.end.y; ++y) {
                 const int mismatches = countMismatches<Mode>(
-                    static_cast<std::int32_t>(x),
-                    static_cast<std::int32_t>(y),
-                    static_cast<std::int32_t>(z),
+                    x,
+                    y,
+                    z,
                     filter.data(),
                     filter.size(),
                     config.errorTolerance);
 
                 if (mismatches <= config.errorTolerance) {
                     matches->push_back({
-                        static_cast<std::int32_t>(x),
-                        static_cast<std::int32_t>(y),
-                        static_cast<std::int32_t>(z),
+                        x,
+                        y,
+                        z,
                         mismatches,
                         item.direction,
                     });
